// src/models/web-annoce.ts
export interface WebAnnoce {

  OrdreNr: number;
  SlutVist: boolean;
  Annoncor: string;
  Betegenelse: string;
  Format: string;
  AntalFarver: number;
  FejlID: number;
  ErOrdre: number;
  Farve: string;
  AnnoncorID: string;
  MmType: number;
  BladID: number;
  DPKulorID: number;
  Version: number;
  MedieplanNr: number;
  EXPR2: number;
  SidePlacering: number;

}